move constants out of own folder, revert to masters package-lock
