+++ .eslintrc
@@ -74,7 +74,7 @@ rules:
     loosely-restrict-imports/loosely-restrict-imports: [2, '.jsx']
     new-cap: 2
     no-cond-assign: 2
-    #no-console: 2
+    no-console: 2
     no-constant-condition: 2
     no-control-regex: 2
     no-debugger: 2
