+++ src/Badge/Badge.js
@@ -1,7 +1,7 @@
 import classnames from 'classnames';
 import PropTypes from 'prop-types';
 import React from 'react';
-import {BADGE_TYPES, LABEL_TYPES, STATUS_TYPES} from '../utils/constants/constants';
+import {BADGE_TYPES, LABEL_TYPES, STATUS_TYPES} from '../utils/constants';
 
 
 export const Badge = ({ type, modifier, children, className, ...props }) => {
