+++ src/InlineHelp/InlineHelp.js
@@ -1,5 +1,5 @@
 import classnames from 'classnames';
-import {INLINE_HELP_PLACEMENTS} from '../utils/constants/constants';
+import {INLINE_HELP_PLACEMENTS} from '../utils/constants';
 import PropTypes from 'prop-types';
 import React from 'react';
