+++ src/Popover/Popover.js
@@ -1,5 +1,5 @@
 import classnames from 'classnames';
-import {POPOVER_ALIGNMENTS} from '../utils/constants/constants';
+import {POPOVER_ALIGNMENTS} from '../utils/constants';
 import PropTypes from 'prop-types';
 import React, { Component } from 'react';
